+++ plugin-packs/postcss-preset-env/src/lib/ids-by-execution-order.mjs
@@ -32,4 +32,5 @@ export default [
 	'overflow-wrap-property',
 	'place-properties',
 	'system-ui-font-family',
+	'cascade-layers',
 ];