+++ plugin-packs/postcss-preset-env/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes to PostCSS Preset Env
 
+### 7.6.0 (May 19, 2022)
+
+- Added `@csstools/postcss-cascade-layers` <br/> [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-cascade-layers#readme) for usage details.
+- Updated `@csstools/postcss-hwb-function`  to `1.0.1` (patch)
+- Updated `@csstools/postcss-is-pseudo-class`  to `2.0.4` (patch)
+- Updated `@csstools/postcss-unset-value`  to `1.0.1` (patch)
+- Updated `postcss-nesting` to `10.1.6` (patch)
+- Updated `postcss-pseudo-class-any-link`  to `7.1.4` (patch)
+
 ### 7.5.0 (May 2, 2022)
 
 - Added `@csstools/postcss-unset-value` <br/> [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-unset-value#readme) for usage details.